third_party/WebKit/LayoutTests/fast/scrolling/scroll-into-view-anchor-next-to-tall-image.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<a name="scrollto">Should Be Visible</a>
<img width="100" height="2000">

<script>
test(function(t) {
  document.querySelector('a').scrollIntoView();
  assert_greater_than(window.scrollY, 1000);
}, "Tests that when the anchor itself has content, scrollIntoView should reveal the anchor regardless of its surrounding elements.");
</script>
